Reworked the whole system to make it more layered.
Reworked the system so the layers are better respected, also the same with solid principles. Made it so that services do not craft the response directly, made response dtos, made the users data stored in spring boot securityContext, reworked the logging for it, made custom exceptions, moved the strings into more appropriate places.