Made it so that now the historical data is upserted, starts from the last candle,  also made the once in a day logic run on the right tickers.
Made it so that now the historical data is upserted, starts from the last candle,  also made the once in a day logic run on the right tickers. The last candle is important, because we might get a not the last for the candle. The upsert logic is there to make this work. The logic for a day now runs when we are not doing historical data synch or do not have a websocket connection.