Made it so that jwt refresh tokens are used accordingly
Made it so that jwt refesh token is being kept track of in the database, when logging in it gets added, when refreshing it also refreshes, when logging out, changing password it gets removed. Also there is a job that would automatically remove it if it expires, it has a smart scheduling system. Also updated the filters to look whether we are looking at a token that is inside out database, also updated the validator for it.