Updated websocket candle return selection logic.
Made it so that the one minute candle now is not handled as atomic, being finished upon creation, instead we also check if that had changed from the last one that we have sent. Also made it so that when there is a new candle we check if the older one has changed and if it has we also send that over.